Edits to auth from class
