+++ solution/29/index.php
@@ -10,7 +10,10 @@
 
 
     // Display the page content
-    $content = render_button('Show Log', 'pagelog.php');
+    // Display the page content
+    $content = render_button('Templates', '../templates');
+    $content .= render_button('Solutions', '..');
+    $content .= render_button('Show Log', 'pagelog.php');
     $content .= $subscribers->handle_actions();
